+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/StashSpell.java
@@ -1,5 +1,6 @@
 package com.elmakers.mine.bukkit.spell.builtin;
 
+import com.zachsthings.libcomponents.Depend;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.HumanEntity;
@@ -12,6 +13,7 @@ import com.elmakers.mine.bukkit.api.magic.Mage;
 import com.elmakers.mine.bukkit.spell.TargetingSpell;
 import com.elmakers.mine.bukkit.utility.Target;
 
+@Deprecated
 public class StashSpell extends TargetingSpell
 {
 	@Override